+++ src/Codeception/Module/Laravel4.php
@@ -25,8 +25,8 @@ use Illuminate\Support\MessageBag;
  *
  * ## Status
  *
- * * Maintainer: **Jon Phipps, Davert**
- * * Stability: **alpha**
+ * * Maintainer: **Davert**
+ * * Stability: **stable**
  * * Contact: davert.codeception@mailican.com
  *
  * ## Config
@@ -36,16 +36,12 @@ use Illuminate\Support\MessageBag;
  * * environment: `string`, default `testing` - When running in unit testing mode, we will set a different environment.
  * * start: `string`, default `bootstrap/start.php` - Relative path to start.php config file.
  * * root: `string`, default ` ` - Root path of our application.
+ * * filters: `boolean`, default: `false` - enable or disable filters for testing.
  *
  * ## API
  *
  * * kernel - `Illuminate\Foundation\Application` instance
  * * client - `BrowserKit` client
- *
- * ## Known Issues
- *
- * When submitting form do not use `Input::all` to pass to store (hope you won't do this anyway).
- * Codeception creates internal form fields, so you get exception trying to save them.
  *
  */
 class Laravel4 extends Framework implements ActiveRecord
@@ -66,6 +62,7 @@ class Laravel4 extends Framework implements ActiveRecord
                 'environment' => 'testing',
                 'start' => 'bootstrap' . DIRECTORY_SEPARATOR . 'start.php',
                 'root' => '',
+                'filters' => false,
             ),
             (array)$config
         );
@@ -79,6 +76,7 @@ class Laravel4 extends Framework implements ActiveRecord
         $this->kernel = $app;
         $this->client = new LaravelConnector($app);
         $this->revertErrorHandler();
+
     }
 
     protected function revertErrorHandler()
@@ -90,13 +88,19 @@ class Laravel4 extends Framework implements ActiveRecord
     public function _before(\Codeception\TestCase $test)
     {
         $this->kernel = $this->getApplication();
+        $this->kernel->boot();
+        $this->kernel->setRequestForConsoleEnvironment();
+
         $this->client = new LaravelConnector($this->kernel);
         $this->client->followRedirects(true);
+
+        if ($this->config['filters']) {
+            $this->haveEnabledFilters();
+        }
 
         if ($this->transactionCleanup()) {
             $this->kernel['db']->beginTransaction();
         }
-//        $this->kernel['router']->enableFilters();
     }
 
     public function _after(\Codeception\TestCase $test)
@@ -118,6 +122,22 @@ class Laravel4 extends Framework implements ActiveRecord
         }
     }
 
+    /**
+     * Enable Laravel filters for next requests.
+     */
+    public function haveEnabledFilters()
+    {
+        $this->kernel['router']->enableFilters();
+    }
+
+    /**
+     * Disable Laravel filters for next requests.
+     */
+    public function haveDisabledFilters()
+    {
+        $this->kernel['router']->disableFilters();
+    }
+
     protected function transactionCleanup()
     {
         return $this->config['cleanup'] and $this->kernel['db'] and $this->expectedLaravelVersion(4.1);
@@ -126,6 +146,76 @@ class Laravel4 extends Framework implements ActiveRecord
     protected function expectedLaravelVersion($ver)
     {
         return floatval(\Illuminate\Foundation\Application::VERSION) >= floatval($ver);
+    }
+
+    /**
+     * Opens web page using route name and parameters.
+     *
+     * ```php
+     * <?php
+     * $I->amOnRoute('posts.create');
+     * ?>
+     * ```
+     *
+     * @param $route
+     * @param array $params
+     */
+    public function amOnRoute($route, $params = [])
+    {
+        $url = $this->kernel['url']->route($route, $params);
+        codecept_debug($url);
+        $this->amOnPage($url);
+    }
+
+    /**
+     * Opens web page by action name
+     *
+     * ```php
+     * <?php
+     * $I->amOnAction('PostsController@index');
+     * ?>
+     * ```
+     *
+     * @param $action
+     * @param array $params
+     */
+    public function amOnAction($action, $params = [])
+    {
+        $url = $this->kernel['url']->action($action, $params);
+        $this->amOnPage($url);
+    }
+
+    /**
+     * Checks that current url matches route
+     *
+     * ```php
+     * <?php
+     * $I->seeCurrentRouteIs('posts.index');
+     * ?>
+     * ```
+     * @param $route
+     * @param array $params
+     */
+    public function seeCurrentRouteIs($route, $params = array())
+    {
+        $this->seeCurrentUrlEquals($this->kernel['url']->route($route, $params, false));
+    }
+
+    /**
+     * Checks that current url matches action
+     *
+     * ```php
+     * <?php
+     * $I->seeCurrentActionIs('PostsController@index');
+     * ?>
+     * ```
+     *
+     * @param $action
+     * @param array $params
+     */
+    public function seeCurrentActionIs($action, $params = array())
+    {
+        $this->seeCurrentUrlEquals($this->kernel['url']->action($action, $params, false));
     }
 
     /**
@@ -204,15 +294,45 @@ class Laravel4 extends Framework implements ActiveRecord
 
     /**
      * Set the currently logged in user for the application.
+     * Takes either `UserInterface` instance or array of credentials.
      *
-     * @param  \Illuminate\Auth\UserInterface $user
+     * @param  \Illuminate\Auth\UserInterface|array $user
      * @param  string $driver
      * @return void
      */
-    public function amLoggedAs(UserInterface $user, $driver = null)
+    public function amLoggedAs($user, $driver = null)
     {
-        $this->kernel['auth']->driver($driver)->setUser($user);
+        if ($user instanceof \Illuminate\Auth\UserInterface) {
+            $this->kernel['auth']->driver($driver)->setUser($user);
+        } else {
+            $this->kernel['auth']->driver($driver)->attempt($user);
+        }
     }
+
+    /**
+     * Logs user out
+     */
+    public function logout()
+    {
+        $this->kernel['auth']->logout();
+    }
+
+    /**
+     * Checks that user is authenticated
+     */
+    public function seeAuthentication()
+    {
+        $this->assertTrue($this->kernel['auth']->check(), 'User is not logged in');
+    }
+
+    /**
+     * Check that user is not authenticated
+     */
+    public function dontSeeAuthentication()
+    {
+        $this->assertFalse($this->kernel['auth']->check(), 'User is logged in');
+    }
+
 
     /**
      * Return an instance of a class from the IoC Container.